# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
    "/api/CarsControllers/GetCars": {
      "get": {
        "tags": [
          "CarsControllers"
        ],
        "operationId": "GetCars",
        "parameters": [
          {
            "name": "color",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocheParaComprarDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocheParaComprarDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocheParaComprarDTO"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/CarsControllers/GetCarsForReview": {
      "get": {
        "tags": [
          "CarsControllers"
        ],
        "operationId": "GetCarsForReview",
        "parameters": [
          {
            "name": "filtroManufacturer",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroFuelType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocheParaReviewDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocheParaReviewDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocheParaReviewDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/CarsControllers/GetCarsRental": {
      "get": {
        "tags": [
          "CarsControllers"
        ],
        "operationId": "GetCarsRental",
        "parameters": [
          {
            "name": "rentingprice",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "modelname",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocheParaAlquilarDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocheParaAlquilarDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocheParaAlquilarDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CocheParaAlquilarDTO": {
        "required": [
          "color",
          "fuelType",
          "manufacturer",
          "modelName",
          "rentingPrice"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "modelName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "color": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string"
          },
          "fuelType": {
            "minLength": 1,
            "type": "string"
          },
          "manufacturer": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "rentingPrice": {
            "maximum": 200000,
            "minimum": 0,
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "CocheParaComprarDTO": {
        "required": [
          "color",
          "fuelType",
          "id",
          "manufacturer",
          "modelName",
          "purchasingPrice"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "modelName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "color": {
            "maxLength": 30,
            "minLength": 0,
            "type": "string"
          },
          "fuelType": {
            "minLength": 1,
            "type": "string"
          },
          "manufacturer": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "purchasingPrice": {
            "maximum": 200000,
            "minimum": 0,
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "CocheParaReviewDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "model": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "carClass": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "manufacturer": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "fuelType": {
            "maxLength": 15,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "color": {
            "maxLength": 20,
            "minLength": 1,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}